// src/components/Items/redux/enums.ts
export enum EItemActions {
	GET_POPULAR_MOVIES_STARTED = 'GET_POPULAR_MOVIES_STARTED',
	GET_POPULAR_MOVIES_SUCCEEDED = 'GET_POPULAR_MOVIES_SUCCEEDED',
	GET_POPULAR_MOVIES_FAILED = 'GET_POPULAR_MOVIES_FAILED',

	GET_POPULAR_SERIES_STARTED = 'GET_POPULAR_SERIES_STARTED',
	GET_POPULAR_SERIES_SUCCEEDED = 'GET_POPULAR_SERIES_SUCCEEDED',
	GET_POPULAR_SERIES_FAILED = 'GET_POPULAR_SERIES_FAILED',

	GET_FAMILY_STARTED = 'GET_FAMILY_STARTED',
	GET_FAMILY_SUCCEEDED = 'GET_FAMILY_SUCCEEDED',
	GET_FAMILY_FAILED = 'GET_FAMILY_FAILED',

	GET_DOCUMENTARIES_STARTED = 'GET_DOCUMENTARIES_STARTED',
	GET_DOCUMENTARIES_SUCCEEDED = 'GET_DOCUMENTARIES_SUCCEEDED',
	GET_DOCUMENTARIES_FAILED = 'GET_DOCUMENTARIES_FAILED',

	GET_MOVIE_STARTED = 'GET_MOVIE_STARTED',
	GET_MOVIE_SUCCEEDED = 'GET_MOVIE_SUCCEEDED',
	GET_MOVIE_FAILED = 'GET_MOVIE_FAILED',

	GET_SERIES_STARTED = 'GET_SERIES_STARTED',
	GET_SERIES_SUCCEEDED = 'GET_SERIES_SUCCEEDED',
	GET_SERIES_FAILED = 'GET_SERIES_FAILED',

	SEARCH_FOR_ITEMS_STARTED = 'SEARCH_FOR_ITEMS_STARTED',
	SEARCH_FOR_ITEMS_SUCCEEDED = 'SEARCH_FOR_ITEMS_SUCCEEDED',
	SEARCH_FOR_ITEMS_FAILED = 'SEARCH_FOR_ITEMS_FAILED'
}